Add mobile menu, reset header font weights
Handles media query that toggles between desktop and mobile menus. Mobile menu slides in well, need to test for usability on touching the hamburger with mobile devices.

KNOWN BUG: When the mobile menu is open, it will slide up proportional to the hiding % of the parent menu when the user scrolls down on the page. To prevent this, the page should either be scroll locked when the mobile menu is open, or the parent header should stay in place. I've tried the latter solution, but for the listener for hiding/showing the menu isn't playing nice with the hook used for the mobile menu toggled class. It may be preferred behavior to just lock the scrolling while open. I'll need to find a way to relay that over to where the HTML tag is though since it is a component scoped variable. Would prefer not to use anything like Redux on this site given its just a side hobby.

Also reset the font weights on the headers as it was clear from the blog posts that 500 weight across the board was not the intended effect of this theme. May experiment with other fonts going forward.