Resolves broken images in markdown blogs
This commit fixes the issue where the salty chinese egg stock image shows up as a broken image. This was happening even though the image in question lived in the same repository as the feature image, which worked fine. Both images were being accessed via a relative path.

The bug was due to a duplicate plugin in the gatsby-config file, with the second one containing much less resolve info than the first.