Add formspree contact form, use custom JS submit and validation logic
With this, Formspree will capture the responses and send me an email with the response. The response itself is not a direct email from the client though.
I considered using node mailer but seems like overkill for a project of this size.

Also decided to center my headers on Contact and Reading List. Might do so on About and might add a header to Writings.

To do: Link to contact page from footer and about